router: Удалил абстракцию Pathr
